requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autocommand==2.2.2
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.csv==1.0.7
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
beautifulsoup4==4.12.2
blinker==1.6.1
blis==0.7.9
Bottleneck @ file:///C:/Windows/Temp/abs_3198ca53-903d-42fd-87b4-03e6d03a8381yfwsuve8/croots/recipe/bottleneck_1657175565403/work
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cheroot==9.0.0
CherryPy==18.8.0
click==8.1.3
codegen==1.0
colorama==0.4.6
confection==0.0.4
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
cymem==2.0.7
datasets==2.10.1
debugpy @ file:///C:/ci/debugpy_1637091961445/work
decorator==5.1.1
dill==0.3.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
entrypoints==0.4
exceptiongroup==1.1.1
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
feedparser==6.0.10
filelock==3.10.0
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
happytransformer==2.4.1
huggingface-hub==0.13.2
idna==3.4
importlib-metadata==6.3.0
importlib-resources==5.12.0
inflect==6.0.4
ipykernel @ file:///D:/bld/ipykernel_1655369313836/work
ipython @ file:///D:/bld/ipython_1677617262387/work
jaraco.collections==4.1.0
jaraco.context==4.3.0
jaraco.functools==3.6.0
jaraco.text==3.11.1
jax==0.4.8
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1676579893731/work
jupyter_core @ file:///D:/bld/jupyter_core_1678994368923/work
keras==2.12.0
kiwisolver==1.4.4
langcodes==3.3.0
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdurl==0.1.2
mkl-service==2.4.0
ml-dtypes==0.1.0
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
mysqlclient==2.1.1
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx==3.0
nltk==3.8.1
numexpr @ file:///C:/b/abs_a7kbak88hk/croot/numexpr_1668713882979/work
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.0
pandas==1.5.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathy==0.10.1
Pattern==3.6
patterns==0.3
pdfminer.six==20221105
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.5.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1678652433955/work
portend==3.1.0
preshed==3.0.8
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1677600924538/work
protobuf==3.20.3
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
py4j @ file:///home/conda/feedstock_root/build_artifacts/py4j_1647581994053/work
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
pyspark @ file:///home/conda/feedstock_root/build_artifacts/pyspark_1676829834027/work
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==305.1
PyYAML==6.0
pyzmq @ file:///C:/ci/pyzmq_1657615952984/work
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.3.3
rsa==4.9
scipy==1.10.1
selenium==4.8.3
semver==2.13.0
sentencepiece==0.1.97
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
st-click-detector==0.1.3
st-clickable-images==0.0.3
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
streamlit==1.20.0
streamlit-image-select==0.6.0
sympy==1.11.1
tempora==5.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
thinc==8.1.9
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==2.0.0
tornado==6.2
tqdm==4.65.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
transformers==4.27.1
trio==0.22.0
trio-websocket==0.10.2
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
wasabi==1.1.1
watchdog==3.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.14.1
wsproto==1.2.0
xxhash==3.2.0
yarl==1.8.2
zc.lockfile==3.0.post1
zipp==3.15.0
